LayoutTests/fast/frames/seamless/seamless-window-location-sandbox-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'about:srcdoc' from frame with URL 'about:srcdoc'. The frame attempting navigation is sandboxed, and is therefore disallowed from navigating its ancestors.

This test should not cause any navigation. 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS 

--------
Frame: '<!--framePath //<!--frame0-->/<!--frame0-->-->'
--------
PASS
